ml-services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
nltk==3.8.1
spacy==3.7.2
requests==2.31.0
aiohttp==3.9.1
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1 